blood_cell_detection/requirements.txt
ultralytics>=8.3.0
roboflow>=1.1.0
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0
PyYAML>=6.0
